first version of a transformer that builds a Journey instance for each valid day of a NetexJourney
unsure about how to handle the days when daylight saving time changes. currently using the "GTFS approach" where all journeys are retained but the times are shifted in a specific way. however there are also netex profiles that recommend to cancel or duplicate specific journeys, but that may not be correct for all operators...